moved GetAssetsRequest to common/models
