+++ alpaca/trading/models/requests.py
@@ -9,9 +9,6 @@ from alpaca.common.enums import (
     Sort,
     TimeInForce,
     OrderClass,
-    AssetStatus,
-    AssetClass,
-    AssetExchange,
 )
 from alpaca.common.models import NonEmptyRequest
 
@@ -259,18 +256,3 @@ class CancelOrderResponse(BaseModel):
 
     id: UUID
     status: int
-
-
-class GetAssetsRequest(NonEmptyRequest):
-    """
-    When querying for available assets, this model provides the parameters that can be filtered by.
-
-    Attributes:
-        status (Optional[AssetStatus]): The active status of the asset.
-        asset_class (Optional[AssetClass]): The type of asset (i.e. us_equity, crypto).
-        exchange (Optional[AssetExchange]): The exchange the asset trades on.
-    """
-
-    status: Optional[AssetStatus]
-    asset_class: Optional[AssetClass]
-    exchange: Optional[AssetExchange]
